tools/telegram_wallet_bot/requirements.txt
python-telegram-bot>=20.0
PyNaCl>=1.5.0
requests>=2.28.0
python-dotenv>=1.0.0
